Fix full path to sound file not being saved
The __on_entries_updated() method updates the mapping of the combo box
items (the name of the file) to their values (the full path) in
pyqtconfig's config manager. When the newly added file is selected
before the mapping is updated, the config manager has no value (full
path) to map the item to, so it just saves the item (name) instead.